# src/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized"}},"summary":"Login user","tags":["Auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully"},"400":{"description":"Bad request"}},"summary":"Register a new user","tags":["Auth"]}},"/users":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User created successfully"},"400":{"description":"Bad request"}},"security":[{"bearer":[]}],"summary":"Create a new user","tags":["Users"]},"get":{"operationId":"UserController_getUsers","parameters":[],"responses":{"200":{"description":"Users found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}},"404":{"description":"Users not found"}},"security":[{"bearer":[]}],"summary":"Get all users","tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UserController_getUser","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"404":{"description":"User not found"}},"security":[{"bearer":[]}],"summary":"Get user by ID","tags":["Users"]},"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"User updated successfully"},"400":{"description":"Bad request"}},"security":[{"bearer":[]}],"summary":"Update user","tags":["Users"]}}},"info":{"title":" API - Users","description":" App","version":"2.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"The email of the user","example":"john.doe@example.com"},"password":{"type":"string","description":"The password of the user","example":"password123"}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the user","example":"John Doe"},"email":{"type":"string","description":"The email of the user","example":"john.doe@example.com"},"password":{"type":"string","description":"The password of the user","example":"password123"}},"required":["name","email","password"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the user","example":"John Doe"},"email":{"type":"string","description":"The email of the user","example":"john.doe@example.com"},"password":{"type":"string","description":"The password of the user","example":"password123"}},"required":["name","email","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"string","description":"The id of the user","example":"c7d7a3c3-3c3c-4c3c-3c3c-3c3c3c3c3c3c"},"name":{"type":"string","description":"The name of the user","example":"John Doe"},"email":{"type":"string","description":"The email of the user","example":"john.doe@example.com"}},"required":["id","name","email"]}}}}